third_party/WebKit/LayoutTests/http/tests/serviceworker/unregister.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>
async_test(function(t) {
  navigator.serviceWorker.unregister('http://example.com/')
    .then(function() {
      assert_unreached('unregistering out of origin should fail');
    }, function(reason) {
      assert_equals(reason.name, 'SecurityError',
                    'unregistering out of origin scope should fail');
      t.done();
    })
    .catch(unreached_rejection(t));
}, 'Unregistering out of origin');

async_test(function(t) {
  navigator.serviceWorker.unregister('/nothing-here/')
    .then(function(value) {
      assert_equals(value, undefined,
                    'unregistering with no existing registration should ' +
                    'succeed with no value');
      t.done();
    })
    .catch(unreached_rejection(t));
}, 'Unregistering with no existing registration');

async_test(function(t) {
  var scope = '/register-then-unregister/';
  navigator.serviceWorker.register(
    'resources/empty-worker.js', {scope: scope})
    .then(function(worker) {
      return navigator.serviceWorker.unregister(scope);
    })
    .then(function(value) {
      assert_equals(value, undefined,
                    'successful unregistration should succeed with no value');
      t.done();
    })
    .catch(unreached_rejection(t));
}, 'Register then unregister');

async_test(function(t) {
  var state_promise;
  service_worker_unregister_and_register(t, 'resources/empty-worker.js')
    .then(function(sw) {
      state_promise = wait_for_state(t, sw, 'redundant');
      return navigator.serviceWorker.unregister();
    })
    .then(function(value) {
      assert_equals(value, undefined,
                    'unregister with default scope should succeed');
      return state_promise;
    })
    .then(function(state) {
      assert_equals(state, 'redundant',
                    'service worker registered with default scope should be ' +
                    'unregstered');
      t.done();
    })
    .catch(unreached_rejection(t));
}, 'Unregistering with default scope');

</script>
